Add files via upload
this is a simple Java code that asks a user to enter numbers, the code then sorts the numbers as either even or odd and then sums up the numbers. the even numbers are added, and the odd numbers are added separately. the sum of both even and odd is also shown.
its outputs are:
1. the sum of even numbers 
2. sum of odd numbers
3. the sum of both even and odd
4. which numbers are even 
5. Which numbers are odd.